+++ api/server.go
@@ -6,12 +6,15 @@ import (
 	"io"
 	"log/slog"
 	"net/http"
+	"regexp"
 	"strconv"
 	"strings"
 	"sync"
 
 	"github.com/anycable/anycable-go/broadcast"
+	"github.com/anycable/anycable-go/broker"
 	"github.com/anycable/anycable-go/server"
+	"github.com/anycable/anycable-go/utils"
 )
 
 // Server manages the HTTP API server
@@ -19,7 +22,9 @@ type Server struct {
 	server *server.HTTPServer
 	conf   *Config
 
-	handler       broadcast.Handler
+	broadcastHandler broadcast.Handler
+	broker           broker.Broker
+
 	authenticator *Authenticator
 
 	enableCORS   bool
@@ -31,7 +36,7 @@ type Server struct {
 }
 
 // NewServer creates a new API server with the given configuration
-func NewServer(conf *Config, handler broadcast.Handler, l *slog.Logger) (*Server, error) {
+func NewServer(conf *Config, brk broker.Broker, handler broadcast.Handler, l *slog.Logger) (*Server, error) {
 	// Derive secret if needed
 	if err := conf.DeriveSecret(); err != nil {
 		return nil, err
@@ -43,10 +48,11 @@ func NewServer(conf *Config, handler broadcast.Handler, l *slog.Logger) (*Server
 	}
 
 	s := &Server{
-		conf:          conf,
-		handler:       handler,
-		authenticator: auth,
-		log:           l.With("context", "api"),
+		conf:             conf,
+		broadcastHandler: handler,
+		broker:           brk,
+		authenticator:    auth,
+		log:              l.With("context", "api"),
 	}
 
 	if conf.AddCORSHeaders {
@@ -78,9 +84,11 @@ func (s *Server) Start() error {
 
 	s.server = srv
 
-	// Register handlers
 	publishPath := s.conf.Path + "/publish"
 	s.server.SetupHandler(publishPath, http.HandlerFunc(s.PublishHandler))
+
+	presencePath := s.conf.Path + "/presence/*"
+	s.server.SetupHandler(presencePath, http.HandlerFunc(s.PresenceHandler))
 
 	return nil
 }
@@ -173,10 +181,54 @@ func (s *Server) PublishHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	if err = s.handler.HandleBroadcast(body); err == nil {
+	if err = s.broadcastHandler.HandleBroadcast(body); err == nil {
 		w.WriteHeader(http.StatusCreated)
 	} else {
 		s.log.Error("failed to handle broadcast", "error", err)
 		w.WriteHeader(http.StatusNotImplemented)
 	}
+}
+
+var channelUsersPathPattern = regexp.MustCompile(`/presence/([^/]+)/users$`)
+
+// PresenseHandler handles GET requests to get the stream's presence info
+func (s *Server) PresenceHandler(w http.ResponseWriter, r *http.Request) {
+	if r.Method != http.MethodGet {
+		s.log.Debug("invalid request method", "method", r.Method)
+		w.WriteHeader(http.StatusUnprocessableEntity)
+		return
+	}
+
+	if !s.authenticator.Authenticate(r) {
+		w.WriteHeader(http.StatusUnauthorized)
+		return
+	}
+
+	matches := channelUsersPathPattern.FindStringSubmatch(r.URL.Path)
+
+	if len(matches) < 2 {
+		s.log.Debug("invalid path for get users", "path", r.URL.Path)
+		w.WriteHeader(http.StatusNotFound)
+		return
+	}
+
+	if !s.broker.SupportsPresence() {
+		w.WriteHeader(http.StatusNotImplemented)
+		return
+	}
+
+	stream := matches[1]
+
+	opts := broker.NewPresenceInfoOptions()
+	info, err := s.broker.PresenceInfo(stream, broker.WithPresenceInfoOptions(opts))
+
+	if err != nil {
+		s.log.Debug("failed to get presence info", "strean", stream, "err", err)
+		w.WriteHeader(http.StatusBadRequest)
+		return
+	}
+
+	w.Header().Set("Content-Type", "application/json")
+	w.WriteHeader(http.StatusOK)
+	w.Write(utils.ToJSON(info)) // nolint:errcheck
 }